fix eslint errors from moving files
